Program that performs simple operations

+++ function_pointers/3-main.c
@@ -24,7 +24,7 @@ int main(int argc, char *argv[])
 		exit(100);
 	}
 	result = (*get_op_func(argv[2]))(atoi(argv[1]), atoi(argv[3]));
-	if (result == NULL)
+	if (result == (int)NULL)
 	{
 		printf("Error\n");
 		exit(99);
